+++ src/tamperbench/whitebox/attacks/lofit/__init__.py
@@ -1,5 +1,6 @@
 """LoFiT: Low-Rank Fine-tuning via Attention Head Selection and Bias Tuning."""
 
 from .lofit_finetune import LoFiTAttack, LoFiTAttackConfig
+from .model_loader import load_lofit_model_and_tokenizer
 
-__all__ = ["LoFiTAttack", "LoFiTAttackConfig"]
+__all__ = ["LoFiTAttack", "LoFiTAttackConfig", "load_lofit_model_and_tokenizer"]
